# api/openapi.yaml
openapi: 3.0.3
info:
  title: Book App API
  version: 1.0.0
  description: |
    REST API for the Book App. The client never connects to the database directly;
    all access is via this API. Resources include profiles, books, follows,
    reviews, reading history, and recommendations.
servers:
  - url: https://api.example.com/v1
    description: Production server (replace with your real base URL)
  - url: http://localhost:3000/v1
    description: Local development server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Review:
      type: object
      properties:
        reviewId:
          type: string
          format: uuid
        body:
          type: string
          nullable: true
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
        worldBuilding:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        characters:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        plot:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        writingStyle:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        reviewDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - reviewId
        - userId
        - bookId
    ReviewCreate:
      type: object
      properties:
        body:
          type: string
          nullable: true
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
        worldBuilding:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        characters:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        plot:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        writingStyle:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
      required:
        - userId
        - bookId
    ReviewUpdate:
      type: object
      properties:
        body:
          type: string
          nullable: true
        worldBuilding:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        characters:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        plot:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        writingStyle:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
...
    Recommendation:
      type: object
      properties:
        recommendationId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
        lastUpdated:
          type: string
          format: date-time
      required:
        - recommendationId
        - userId
        - bookId
    RecommendationCreate:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
      required:
        - userId
        - bookId
    RecommendationUpdate:
      type: object
      properties:
        lastUpdated:
          type: string
          format: date-time
...
paths:
...
  /reviews:
    get:
      tags: [Reviews]
      summary: List reviews
      operationId: listReviews
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: bookId
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
    post:
      tags: [Reviews]
      summary: Create a review
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid input
  /reviews/{reviewId}:
    get:
      tags: [Reviews]
      summary: Get review by ID
      operationId: getReviewById
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Review not found
    patch:
      tags: [Reviews]
      summary: Update review
      operationId: updateReview
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdate'
      responses:
        '200':
          description: Updated review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Review not found
    delete:
      tags: [Reviews]
      summary: Delete review
      operationId: deleteReview
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Review deleted
        '404':
          description: Review not found
...
  /recommendations:
    get:
      tags: [Recommendations]
      summary: List recommendations
      operationId: listRecommendations
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recommendation'
    post:
      tags: [Recommendations]
      summary: Create a recommendation
      operationId: createRecommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationCreate'
      responses:
        '201':
          description: Recommendation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '400':
          description: Invalid input
        '409':
          description: Recommendation already exists for user+book
  /recommendations/{id}:
    get:
      tags: [Recommendations]
      summary: Get recommendation by ID
      operationId: getRecommendationById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recommendation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '404':
          description: Recommendation not found
    patch:
      tags: [Recommendations]
      summary: Update recommendation
      operationId: updateRecommendation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationUpdate'
      responses:
        '200':
          description: Updated recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '404':
          description: Recommendation not found
    delete:
      tags: [Recommendations]
      summary: Delete recommendation
      operationId: deleteRecommendation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recommendation deleted
        '404':
          description: Recommendation not found